requirements.txt
discord.py==2.3.2
openai==1.69.0
python-dotenv==1.1.0
Requests==2.32.3
selenium==4.30.0
webdriver_manager==4.0.2
